eddyProStats loads and plots the "stats" output from EddyPro, notably the cov(uw), aka another way of calculating the momentum flux.
hydroDynamicsVanniBuoy loads the temperature profile data from the Vanni group's buoy and uses RLakeAnalyzer functions to calculate W, make plots.

loadCospectraHarsha is copy+paste of LoadCospectra, changed to point to the directories where the Harsha Lake drawdown results are housed. Loads and plots the cospectra from the week of observations.